refactor: extrenal in-memory cache handling moved to the responsible class
