+++ README.md
@@ -49,7 +49,7 @@ The external resource parsers can be defined using the `parsers` attribute of th
     parsers: {
         rawInput?: (data: string) => unknown; // Raw input string -> data (e.g. JSON.parse)
         input?: BaseSchema || (data: unknown) => unknown; // data intake -> parsed data schema
-        transform?: (data: unknown, params: Record<string, string>) => Promise<unknown>; // Parsed data intake with query params obj -> transformed data
+        transform?: (data: unknown, params?: Record<string, string>) => Promise<unknown>; // Parsed data intake with query params obj -> transformed data
         output?: BaseSchema || (data: unknown) => unknown; // Transformed data intake -> output data schema
         rawOutput?: (data: unknown) => string; // Output data intake -> raw output string (e.g. JSON.stringify)
     }
