+++ src/utils/data/repositories/ResourceRepository.ts
@@ -1,5 +1,5 @@
 import * as resources from '../../../resources/index';
-import { IResource } from '../models/shared/BaseResource';
+import IResource from '../models/shared/IResource';
 const _resourses = resources.default;
 
 export function getResources(resourceType: 'external' | 'internal' | 'library' | 'all' = 'all') {
